+++ examples/examplecode/data/example_data_multilabel.py
@@ -1,5 +1,4 @@
 from sklearn.feature_extraction.text import TfidfVectorizer
-from sklearn.preprocessing import normalize
 
 from small_text.data import SklearnDataset
 from small_text.utils.labels import list_to_csr
@@ -18,10 +17,10 @@ def preprocess_data_sklearn(train, test):
 
     vectorizer = TfidfVectorizer(stop_words='english')
 
-    x_train = normalize(vectorizer.fit_transform(train['text']))
-    x_test = normalize(vectorizer.transform(test['text']))
-
     y_train = list_to_csr(train['labels'], shape=(len(train), NUM_LABELS))
     y_test = list_to_csr(test['labels'], shape=(len(test), NUM_LABELS))
 
-    return SklearnDataset(x_train, y_train), SklearnDataset(x_test, y_test)
+    ds_train = SklearnDataset.from_arrays(train['text'], y_train, vectorizer, train=True)
+    ds_test = SklearnDataset.from_arrays(test['text'], y_test, vectorizer, train=False)
+
+    return ds_train, ds_test