+++ small_text/data/__init__.py
@@ -1,2 +1,16 @@
 from small_text.data.datasets import Dataset, DatasetView, SklearnDataset
-from small_text.data.sampling import stratified_sampling, balanced_sampling
+from small_text.data.sampling import (
+    balanced_sampling,
+    multilabel_stratified_subsets_sampling,
+    stratified_sampling
+)
+
+
+__all__ = [
+    'Dataset',
+    'DatasetView',
+    'SklearnDataset',
+    'balanced_sampling',
+    'multilabel_stratified_subsets_sampling',
+    'stratified_sampling'
+]